Refactor out types
