+++ src/market/__init__.py
@@ -1,2 +1,2 @@
-from .market import Market, Order, OrderBook, OrderInfo  # noqa: F401
-from .state import AccountFlags, MarketState  # noqa: F401
+from .market import Market, OrderBook  # noqa: F401
+from .state import MarketState as State  # noqa: F401
